StayPreview to show .0 when rating is round. moved func to util.service.js
